Adjust tests to a more realistic expectation
Previously I expected a cache SET metric to trigger when a cache GetOrSetAsync's function threw an exception.   After the backplane feature was added in FusionCache this flaw in logic was exposed.  I think you can logically draw the parallel conclusion that when failsafe feature was not enabled it does not raise an event indicating cache was set.  So, I am updating the test to match what is the more correct behavior.