Starting the OpenTelemetry work.
Will turn it into a full OpenTelemetry example project with metrics, traces and logs!

The goal of this feature will be to include a OpenTelemetry plugin for FusionCache.  An OpenTelemetry plugin was the original goal for this repo but it wasn't ready enough in the past.  Thus we created a EventCounters plugin and an AppMetrics open source  plugin for the time being.