Stop using GitHub prerelease and publish during build
The $GITHUB_RUN_ID variable seems to be set to a different style number and is breaking my basic build.  I am not using those prerelease builds anyway.  I do have a prerelease technique via tagging and activities that react to the type of tag and publish on Nuget.
One test is flaky on build server.  Adding a second longer delay.