Upgrade FusionCache to 0.16
This branch has some OpenTelemetery work in it.  This will eventually get published in the OpenTelemetry Contrib project.  For now it was more up to date and I am behind on keeping the Develop branch up to date.  So there will be more commits from this branch to update packages.